Finished updating project code style. Rewrote Chip8 interpreter to use functions for ops. Rewrote Chiup8Test to unit test op functions. Fixed bug in op DRW where the flag register was not cleared before drawing.

+++ src/KeyHandler.hpp
@@ -16,7 +16,7 @@ namespace Chip8{
 
         struct KeyPair{
             SDL_Keycode key;
-            uint16_t modifier;
+            int32_t modifier;
 
             bool operator==(const KeyPair& t_other) const{
                 return key == t_other.key && (modifier == t_other.modifier || (modifier & t_other.modifier));
